+++ src/background/DappService.js
@@ -1,8 +1,8 @@
 import { DAppActions } from '@aurowallet/mina-provider';
-import extension from 'extensionizer';
+import browser from 'webextension-polyfill';
 import ObservableStore from "obs-store";
 import { DAPP_ACTION_CANCEL_ALL, DAPP_ACTION_CLOSE_WINDOW, DAPP_ACTION_CREATE_NULLIFIER, DAPP_ACTION_GET_ACCOUNT, DAPP_ACTION_REQUEST_PRESENTATION, DAPP_ACTION_SEND_TRANSACTION, DAPP_ACTION_SIGN_MESSAGE, DAPP_ACTION_STORE_CREDENTIAL, DAPP_ACTION_SWITCH_CHAIN, WORKER_ACTIONS } from '../constant/msgTypes';
-import { checkAndTop, checkAndTopV2, closePopupWindow, lastWindowIds, openPopupWindow, PopupSize, startExtensionPopup, startPopupWindow } from "../utils/popup";
+import { checkAndTopV2, closePopupWindow, lastWindowIds, PopupSize, startExtensionPopup, startPopupWindow } from "../utils/popup";
 import { checkNodeExist, getArrayDiff, getCurrentNodeConfig, getExtensionAction, getLocalNetworkList, getMessageFromCode, getOriginFromUrl, isNumber, removeUrlFromArrays, urlValid } from '../utils/utils';
 import { addressValid } from '../utils/validator';
 import apiService from './APIService';
@@ -335,7 +335,7 @@ class DappService {
             that.setBadgeContent()
             if(chainRequests.length === 0){
               closePopupWindow(windowId.request_sign) 
-              extension.runtime.onMessage.removeListener(onMessage)
+              browser.runtime.onMessage.removeListener(onMessage)
               that.signEventListener = undefined
             }
             return 
@@ -358,7 +358,7 @@ class DappService {
                   nextReject({code: errorCodes.userRejectedRequest , message:getMessageFromCode(errorCodes.userRejectedRequest)})
                   that.removeSignParamsByOpenId(payload.id)
                   if(signRequests.length == 0 && chainRequests.length ===0){
-                    extension.runtime.onMessage.removeListener(onMessage)
+                    browser.runtime.onMessage.removeListener(onMessage)
                     that.signEventListener = undefined
                   }
                   that.setBadgeContent()
@@ -368,7 +368,7 @@ class DappService {
                   })
                   that.removeSignParamsByOpenId(payload.id)
                   if(signRequests.length == 0 && chainRequests.length ===0){
-                    extension.runtime.onMessage.removeListener(onMessage)
+                    browser.runtime.onMessage.removeListener(onMessage)
                     that.signEventListener = undefined
                   }
                   that.setBadgeContent()
@@ -387,7 +387,7 @@ class DappService {
                   nextReject({code: errorCodes.userRejectedRequest , message:getMessageFromCode(errorCodes.userRejectedRequest)})
                   that.removeSignParamsByOpenId(payload.id)
                   if(signRequests.length == 0 && chainRequests.length ===0){
-                    extension.runtime.onMessage.removeListener(onMessage)
+                    browser.runtime.onMessage.removeListener(onMessage)
                     that.signEventListener = undefined
                   }
                   that.setBadgeContent()
@@ -397,7 +397,7 @@ class DappService {
                   })
                   that.removeSignParamsByOpenId(payload.id)
                   if(signRequests.length == 0 && chainRequests.length ===0){
-                    extension.runtime.onMessage.removeListener(onMessage)
+                    browser.runtime.onMessage.removeListener(onMessage)
                     that.signEventListener = undefined
                   }
                   that.setBadgeContent()
@@ -425,7 +425,7 @@ class DappService {
                     }
                     that.removeSignParamsByOpenId(payload.id)
                     if(signRequests.length == 0 && chainRequests.length ===0){
-                      extension.runtime.onMessage.removeListener(onMessage)
+                      browser.runtime.onMessage.removeListener(onMessage)
                       that.signEventListener = undefined
                     }
                     that.setBadgeContent()
@@ -433,7 +433,7 @@ class DappService {
                     nextReject({code: errorCodes.userRejectedRequest , message:getMessageFromCode(errorCodes.userRejectedRequest)})
                     that.removeSignParamsByOpenId(payload.id)
                     if(signRequests.length == 0 && chainRequests.length ===0){
-                      extension.runtime.onMessage.removeListener(onMessage)
+                      browser.runtime.onMessage.removeListener(onMessage)
                       that.signEventListener = undefined
                     }
                     that.setBadgeContent()
@@ -452,7 +452,7 @@ class DappService {
                     nextResolve(payload)
                     that.removeSignParamsByOpenId(payload.id)
                     if(signRequests.length == 0 && chainRequests.length ===0){
-                      extension.runtime.onMessage.removeListener(onMessage)
+                      browser.runtime.onMessage.removeListener(onMessage)
                       that.signEventListener = undefined
                     }
                     that.setBadgeContent()
@@ -462,7 +462,7 @@ class DappService {
                     nextReject({code: errorCodes.userRejectedRequest , message:getMessageFromCode(errorCodes.userRejectedRequest)})
                     that.removeSignParamsByOpenId(payload.id)
                     if(signRequests.length == 0 && chainRequests.length ===0){
-                      extension.runtime.onMessage.removeListener(onMessage)
+                      browser.runtime.onMessage.removeListener(onMessage)
                       that.signEventListener = undefined
                     }
                     that.setBadgeContent()
@@ -479,7 +479,7 @@ class DappService {
                     nextReject({code: errorCodes.userRejectedRequest , message:getMessageFromCode(errorCodes.userRejectedRequest)})
                     that.removeNotifyParamsByOpenId(payload.id)
                     if(signRequests.length == 0 && chainRequests.length ===0){
-                      extension.runtime.onMessage.removeListener(onMessage)
+                      browser.runtime.onMessage.removeListener(onMessage)
                       that.signEventListener = undefined
                     }
                     that.setBadgeContent()
@@ -487,7 +487,7 @@ class DappService {
                     nextResolve(payload.nextConfig)
                     that.removeNotifyParamsByOpenId(payload.id)
                     if(signRequests.length == 0 && chainRequests.length ===0){
-                      extension.runtime.onMessage.removeListener(onMessage)
+                      browser.runtime.onMessage.removeListener(onMessage)
                       that.signEventListener = undefined
                     }
                     that.setBadgeContent()
@@ -507,7 +507,7 @@ class DappService {
                   nextResolve(payload)
                   that.removeSignParamsByOpenId(payload.id)
                   if(signRequests.length == 0 && chainRequests.length ===0){
-                    extension.runtime.onMessage.removeListener(onMessage)
+                    browser.runtime.onMessage.removeListener(onMessage)
                     that.signEventListener = undefined
                   }
                   that.setBadgeContent()
@@ -517,7 +517,7 @@ class DappService {
                   nextReject({code: errorCodes.userRejectedRequest , message:getMessageFromCode(errorCodes.userRejectedRequest)})
                   that.removeSignParamsByOpenId(payload.id)
                   if(signRequests.length == 0 && chainRequests.length ===0){
-                    extension.runtime.onMessage.removeListener(onMessage)
+                    browser.runtime.onMessage.removeListener(onMessage)
                     that.signEventListener = undefined
                   }
                   that.setBadgeContent()
@@ -528,7 +528,7 @@ class DappService {
           return false
         }
         if(!that.signEventListener){
-          that.signEventListener = extension.runtime.onMessage.addListener(onMessage)
+          that.signEventListener = browser.runtime.onMessage.addListener(onMessage)
         }
         let time = new Date().getTime()
         if(ZKAPP_CHAIN_ACTION.indexOf(sendAction)!==-1){
@@ -658,7 +658,7 @@ class DappService {
                   nextReject({ message: getMessageFromCode(errorCodes.originDismatch),code:errorCodes.originDismatch })
                   return
                 }
-                extension.runtime.onMessage.removeListener(onMessage)
+                browser.runtime.onMessage.removeListener(onMessage)
                 nextResolve([payload.account])
                 approveRequests = []
                 that.setBadgeContent()
@@ -669,7 +669,7 @@ class DappService {
           }
           return false
         }
-        extension.runtime.onMessage.addListener(onMessage)
+        browser.runtime.onMessage.addListener(onMessage)
         approveRequests.push({ id, site,resolve,reject })
         this.setBadgeContent()
         sendMsg({
@@ -860,24 +860,25 @@ class DappService {
     return
   }
   getTabPermission (){
-    return new Promise(()=>{
-      chrome.permissions.contains(
+    return new Promise((resolve)=>{
+      browser.permissions.contains(
         {
           permissions: ["tabs"],
         },
-        (result) => {
-          isHaveTabPermissions = result;
-          return result
-        }
-      );
+      ).then((result)=>{
+        isHaveTabPermissions = result;
+        return result
+      }).finally(()=>{
+        resolve()
+      });
     })
   }
  async notifyAccountChange(siteUrlList, connectAccount) {
     if(!isHaveTabPermissions){
-      await getTabPermission()
+      await this.getTabPermission()
     }
     let account = !connectAccount ? [] : [connectAccount]
-    extension.tabs.query({}, (tabs) => {
+    browser.tabs.query({}).then((tabs)=>{
       let message = {
         action: "accountsChanged",
         result: account
@@ -889,10 +890,10 @@ class DappService {
 
             let tabConnectIndex = siteUrlList.indexOf(origin)
             if(tabConnectIndex !== -1){
-              extension.tabs.sendMessage(tab.id, message, res => { })
+              browser.tabs.sendMessage(tab.id, message)
             }
           }else{
-            extension.tabs.sendMessage(tab.id, message, res => { })
+            browser.tabs.sendMessage(tab.id, message)
           }
           continue;
       }
@@ -900,7 +901,7 @@ class DappService {
   }
   async notifyNetworkChange(currentNet) {
     if(!isHaveTabPermissions){
-      await getTabPermission()
+      await this.getTabPermission()
     }
     let networkID = currentNet.networkID || ""
     let message = {
@@ -912,19 +913,18 @@ class DappService {
     this.tabNotify(message)
   }
   tabNotify(message){
-    extension.tabs.query({
-    }, (tabs) => {
-      let currentConnect = this.getDappStore().currentConnect
-      for (let index = 0; index < tabs.length; index++) {
-        const tab = tabs[index];
-        if(isHaveTabPermissions){
-          if (currentConnect[tab.id]) {
-            extension.tabs.sendMessage(tab.id, message, res => { })
+    browser.tabs.query({}).then((tabs)=>{
+        let currentConnect = this.getDappStore().currentConnect
+        for (let index = 0; index < tabs.length; index++) {
+          const tab = tabs[index];
+          if(isHaveTabPermissions){
+            if (currentConnect[tab.id]) {
+              browser.tabs.sendMessage(tab.id, message)
+            }
+          }else{
+            browser.tabs.sendMessage(tab.id, message)
           }
-        }else{
-          extension.tabs.sendMessage(tab.id, message, res => { })
         }
-      }
     })
   }
   portDisconnectListener(port) {
@@ -1023,7 +1023,7 @@ class DappService {
         top: buildParams.top  + PopupSize.exitSize,
       }
     }
-    startPopupWindow(targetUrl, "tokenSign_"+buildID, "buildDapp", {
+    startPopupWindow(targetUrl, "tokenSign_"+buildID, {
       ...nextOption
     });
     return buildID
@@ -1119,7 +1119,7 @@ class DappService {
             that.setBadgeContent()
             if(tokenSigneRequests.length === 0){
               closePopupWindow(windowId.token_sign) 
-              extension.runtime.onMessage.removeListener(onMessage)
+              browser.runtime.onMessage.removeListener(onMessage)
               that.tokenSignListener = undefined
             }
             return 
@@ -1154,7 +1154,7 @@ class DappService {
                     that.removeTokenBuildById(payload.id)
                     if(tokenSigneRequests.length == 0){
                       closePopupWindow(windowId.token_sign)
-                      extension.runtime.onMessage.removeListener(onMessage)
+                      browser.runtime.onMessage.removeListener(onMessage)
                       that.tokenSignListener = undefined
                     }
                     that.setBadgeContent()
@@ -1163,7 +1163,7 @@ class DappService {
                     that.removeTokenBuildById(payload.id)
                     if(tokenSigneRequests.length == 0 ){
                       closePopupWindow(windowId.token_sign)
-                      extension.runtime.onMessage.removeListener(onMessage)
+                      browser.runtime.onMessage.removeListener(onMessage)
                       that.tokenSignListener = undefined
                     }
                     that.setBadgeContent()
@@ -1193,7 +1193,7 @@ class DappService {
         nextParams.buildData = buildData;
         nextParams.result = decryptData.transaction
         if(!that.tokenSignListener){ 
-          that.tokenSignListener = extension.runtime.onMessage.addListener(onMessage)
+          that.tokenSignListener = browser.runtime.onMessage.addListener(onMessage)
         }
         let time = new Date().getTime()
         tokenSigneRequests.push({ id, params:nextParams, site,resolve,reject,time })