+++ src/background/extensionStorage.js
@@ -1,10 +1,10 @@
-import extension from "extensionizer";
+import browser from 'webextension-polyfill';
 
 /**
  * save local
  */
  export async function extSaveLocal(key, value) {
-    await extension.storage.local.set({[key]: value});
+    await browser.storage.local.set({[key]: value});
 }
 
 
@@ -13,7 +13,7 @@ import extension from "extensionizer";
  * @param {*} key
  */
 export async function extGetLocal(key) {
-    const data = await extension.storage.local.get([key]);
+    const data = await browser.storage.local.get([key]);
     return data[key]
 }
 
@@ -22,38 +22,5 @@ export async function extGetLocal(key) {
  * @param {*} key
  */
 export async function extRemoveLocal(key) {
-  return extension.storage.local.remove(key);
-}
-
-/**
- * remove all local storage
- */
-export async function extClearLocal() {
-  return extension.storage.local.clear();
-}
-
-/**
- * get all local storage
- */
- export async function extGetAllLocal() {
-    return extension.storage.local.get();
-}
-
-
-/**
- * clear local storage except some key
- */
- export async function clearLocalExcept(targetKey) {
-     if(targetKey){
-        let data = await extGetAllLocal()
-        let keys = Object.keys(data)
-        for (let index = 0; index < keys.length; index++) {
-            const key = keys[index];
-            if(key!== targetKey ){
-                await extRemoveLocal(key)
-            }
-        }
-     }else{
-        await extClearLocal()
-     }
+  return browser.storage.local.remove(key);
 }