+++ src/popup/pages/index.js
@@ -1,23 +1,22 @@
-import extension from "extensionizer";
-import React, { useCallback, useEffect, useState } from "react";
+import { useCallback, useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useHistory } from "react-router-dom";
+import styled, { keyframes } from "styled-components";
+import browser from "webextension-polyfill";
 import { getBaseInfo } from "../../background/api";
-import { FROM_BACK_TO_RECORD, SET_LOCK, WORKER_ACTIONS } from "../../constant/msgTypes";
+import { FROM_BACK_TO_RECORD, WORKER_ACTIONS } from "../../constant/msgTypes";
 import { languageInit } from "../../i18n";
 import { setLanguage } from "../../reducers/appReducer";
-import { updateExtensionBaseInfo, updatePopupLockStatus } from "../../reducers/cache";
+import {
+  updateExtensionBaseInfo,
+  updatePopupLockStatus,
+} from "../../reducers/cache";
 import {
   ENTRY_WITCH_ROUTE,
   updateEntryWitchRoute,
 } from "../../reducers/entryRouteReducer";
-import ApprovePage from "./ApprovePage";
 import { LockPage } from "./Lock";
 import HomePage from "./Main";
-import SignTransaction from "./SignTransaction";
-import Welcome from "./Welcome";
-import styled, { keyframes } from "styled-components";
-import TokenSignPage from "./Send/tokenSign";
 
 const MainRouter = () => {
   const dispatch = useDispatch();
@@ -31,7 +30,7 @@ const MainRouter = () => {
   const initBaseInfo = useCallback(async () => {
     let baseInfo = await getBaseInfo().catch((err) => err);
     if (baseInfo) {
-      dispatch(updateExtensionBaseInfo(baseInfo)); 
+      dispatch(updateExtensionBaseInfo(baseInfo));
     }
   }, []);
 
@@ -45,25 +44,20 @@ const MainRouter = () => {
     let lockEvent = (message, sender, sendResponse) => {
       const { type, action, payload } = message;
       if (type === FROM_BACK_TO_RECORD && action === WORKER_ACTIONS.SET_LOCK) {
-        if(!payload){
+        if (!payload) {
           dispatch(updateEntryWitchRoute(ENTRY_WITCH_ROUTE.LOCK_PAGE));
           history.push("/lock_page");
         }
-        dispatch(updatePopupLockStatus(!payload))
-        sendResponse()
+        dispatch(updatePopupLockStatus(!payload));
+        sendResponse();
       }
       return false;
     };
-    extension.runtime.onMessage.addListener(lockEvent);
+    browser.runtime.onMessage.addListener(lockEvent);
   }, []);
 
   useEffect(() => {
     switch (entryWitchRoute) {
-      case ENTRY_WITCH_ROUTE.WELCOME:
-        extension.tabs.create({
-          url: "popup.html#/register_page",
-        });
-        return;
       case ENTRY_WITCH_ROUTE.HOME_PAGE:
         setNextRoute(<HomePage />);
         return;
@@ -109,4 +103,3 @@ const LoadingView = () => {
     </StyledLoadingWrapper>
   );
 };
-