+++ CHANGELOG.md
@@ -7,9 +7,11 @@ All notable changes to this project will be documented in this file.
 - Credentials icon
 - Add support for revoke permissions
 - Add permission `tabs` to fix [#50](https://github.com/aurowallet/auro-wallet-browser-extension/issues/50)
+- Update `extensionizer` to `webextension-polyfill`
 
 ### Fixes
 - Remove storybook
+- Legder confirm button ui
 
 ## [2.3.3]
 ### Fixes